requirements.txt
beautifulsoup4==4.13.4
chromadb==0.6.3
gradio==5.25.2
langchain==0.3.23
langchain_community==0.3.21
python_docx==1.1.2
Requests==2.32.3
sentence_transformers==4.1.0
